// app/components/WordpressStat/interfaces/wordpress-stat-props.interface.ts
export interface WordpressStatPropsInterface {
  mainCaption: string;
  dateCaption: string;
  mainMbActive?: boolean;
  mainMbQuantityCaption?: string;
  thisMonthActive?: boolean;
  thisMonth?: string;
  dayQuantity: string;
  diskUsageCaption?: string;
  mbQuantity?: string;
  uniqueVisit?: number;
  bluePoint?: boolean;
}